Set up container for Fedora development
The Containerfile creates a container that includes required RPMs, and sets up
an account with appropriate configuration to perform Fedora development using
tools such as `fedpkg` and Koji.

This container image can be useful when using, for instance, Podman on Mac, to
perform Fedora development tasks.

A Makefile has shortcuts to the relevant commands to build and run the container.